# Conflicts:
#	dependencies/libs/dependencies.lib
#	gradle.properties
#	harbor-crypto/src/main/java/com/ziyao/harbor/crypto/PropertyResolver.java
#	oauth2/oauth2-core/src/main/java/com/ziyao/security/oauth2/core/error/OAuth2AuthenticationException.java
#	springframework/autoconfigure/src/main/java/com/harbor/boot/autoconfigure/servlet/WebMvcAutoConfiguration.java
#	springframework/autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
#	springframework/spring-data-redis/src/main/java/com/ziyao/harbor/data/redis/core/CommonOperations.java
#	springframework/spring-oauth2-core/src/main/java/com/ziyao/oauth2/core/DefaultOAuth2AccessToken.java
#	springframework/spring-oauth2-core/src/main/java/com/ziyao/oauth2/core/TokenType.java
#	springframework/spring-oauth2-core/src/main/java/com/ziyao/oauth2/error/OAuth2Exception.java
#	springframework/spring-security-oauth2/spring-security-oauth2.gradle
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/error/AccessTokenRequiredException.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/error/OAuth2Error.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/http/SslHttpRequestFactory.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/http/converter/HttpMessageConverters.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/provider/authentication/BearerTokenExtractor.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/provider/authentication/OAuth2AuthenticationDetails.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/provider/authentication/OAuth2AuthenticationDetailsSource.java
#	springframework/spring-security-oauth2/src/main/java/com/ziyao/harbor/oauth2/provider/authentication/TokenExtractor.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/UserCenterStarter.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/authenticate/provider/PasswordAuthenticator.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/authenticate/query/UserQuery.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/authenticate/support/PasswordParameter.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/authenticate/support/PasswordValidator.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/authentication/codec/BCryptPasswordEncryptor.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/comm/exception/AbstractException.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/comm/exception/AuthenticateException.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/comm/exception/UserStatusException.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/dto/AppDTO.java
#	user-center/src/main/java/com/ziyao/harbor/usercenter/entity/App.java

+++ user-center/src/main/java/com/ziyao/harbor/usercenter/dto/MenuDTO.java
@@ -17,7 +17,7 @@ import java.util.Objects;
  * </p>
  *
  * @author zhangziyao
- * @since 2023-05-09
+ * @since 2024-06-08
  */
 @Data
 public class MenuDTO implements EntityDTO<Menu>, Serializable {
@@ -56,7 +56,7 @@ public class MenuDTO implements EntityDTO<Menu>, Serializable {
     /**
      * 资源级别
      */
-    private Integer level;
+    private Byte level;
     /**
      * 排序
      */
